chore: tweak dir structure
