+++ packages/global/common/tsRest/fastgpt/contracts/chat/index.ts
@@ -1,0 +1,6 @@
+import { settingContract } from './setting';
+import { c } from '../../../init';
+
+export const chatContract = c.router({
+  setting: settingContract
+});
