--- packages/global/common/tsRest/fastgptpro/contracts/chat/index.ts
@@ -1,6 +1,0 @@
-import { settingProContract } from './setting';
-import { c } from '../../../init';
-
-export const chatProContract = c.router({
-  setting: settingProContract
-});
